On branch master  Your branch is up-to-date with 'origin/master'. Added kernel images and Device config form Samsung Galaxy note 3 europe aka hlte. Noted issues:
TERMINAL:
Terminal all built with the package doesn't work out of the box
needed to be replaced by
https://github.com/offensive-security/kali-nethunter/blob/master/nethunter-installer/update/data/app/Term-nh.apk
Confirmed working with the version form the link above.

cSploit: Version built with the package, missing a libcSploitCommon.so
Confirmed working with the version, last release:
https://github.com/cSploit/android/releases

routerkeygen: Version built with package also not working.
Confirmed working with the following version:

https://github.com/routerkeygen/routerkeygenAndroid/releases/download/v3.15.0/routerkeygen-3-15-0.apk

 Changes to be committed:
	modified:   devices.cfg
	new file:   nougat/hlteeur-cm/zImage
	new file:   nougat/hlteeur-cm/zImage-dtb